requirements.txt
python-telegram-bot>=20.7,<22.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.1
yt-dlp>=2025.3.31
openai-whisper>=20240930
python-dotenv>=1.0.1
groq>=0.18.0
